requirements.txt
airflow-provider-great-expectations==0.0.6
apache-airflow-providers-snowflake==2.1.0
azure-storage-blob==12.8.1
certifi==2021.5.30